requirements.txt
fastapi
uvicorn
requests
python-dotenv
ngrok
